Added DELETE endpoint

+++ src/main/java/javau9/pizzeria/services/PizzaServiceImpl.java
@@ -9,33 +9,55 @@ import javau9.pizzeria.repositories.PizzaRepository;
 import javau9.pizzeria.models.Pizza;
 
 @Service
-public class PizzaServiceImpl implements PizzaService{
+public class PizzaServiceImpl implements PizzaService {
 
-    PizzaRepository pizzaDatabase;
-    
-    public PizzaServiceImpl(PizzaRepository pizzaDatabase) {
-    	this.pizzaDatabase = pizzaDatabase;
-    }
-       
-    public Pizza addPizza(Pizza pizza) {
-        return pizzaDatabase.save(pizza);
-    }
+	PizzaRepository pizzaDatabase;
 
-    public Optional<Pizza> getPizzaById(Long id) {
-    	System.out.println( "-----> " + id );
-        return pizzaDatabase.findById(id);
-    }
+	public PizzaServiceImpl(PizzaRepository pizzaDatabase) {
+		this.pizzaDatabase = pizzaDatabase;
+	}
 
-    public Collection<Pizza> getAllPizzas() {
-        return pizzaDatabase.findAll();
-    }
+	public Pizza addPizza(Pizza pizza) {
+		return pizzaDatabase.save(pizza);
+	}
 
-    public void removePizza(Long id) {
-        pizzaDatabase.deleteById(id);
-    }
+	public Optional<Pizza> getPizzaById(Long id) {
+		System.out.println("-----> " + id);
+		return pizzaDatabase.findById(id);
+	}
+
+	public Collection<Pizza> getAllPizzas() {
+		return pizzaDatabase.findAll();
+	}
+
+	public void removePizza(Long id) {
+		pizzaDatabase.deleteById(id);
+	}
 
 	@Override
 	public Optional<Pizza> updatePizza(Long id, Pizza pizza) {
-		return Optional.of(  pizzaDatabase.save(pizza) );
+		
+		Optional<Pizza> existingPizza = pizzaDatabase.findById(id);
+		if (existingPizza.isEmpty())
+			return Optional.empty();
+
+		Pizza updatedPizza = existingPizza.get();
+		updatedPizza.setName(	pizza.getName() );
+		updatedPizza.setSize(	pizza.getSize() );
+		updatedPizza.setPrice(	pizza.getPrice() );
+
+		return Optional.of(pizzaDatabase.save(updatedPizza));
+
 	}
+
+	@Override
+    public boolean deletePizza(Long id) {
+        if (pizzaDatabase.existsById(id)) {
+        	pizzaDatabase.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+	
+	
 }
